+++ src/Parser/Env.hs
@@ -1,5 +1,0 @@
-module Parser.Env
-  ( module Env
-  ) where
-
-import Env